Move `FeedStore` & `LocaleFeedLoader` to production in their own files
